Create the class Alien to model an alien in the game
Not having a central code block to modify to change the aliens in the game would be bad.

It would be challenging to change and maintain the aliens' details and behavior.

Prevent the above from happening by creating a class Alien to model an alien in the game.

Create a method to display the alien on the screen.