Refactor create_fleet function
Creating a new row or fleet will lengthen the code blocks in the original function.

Making it difficult to maintain and modify.

It is necessary to refactor the original function to allow for better implementation.

Create function get_number_aliens_x to get the number of aliens that will fit in a row.

Create function create_alien to create an alien and place it in the row.

Update function create_fleet.