Add functionality for the alien fleet to move
The alien fleet is static.

Enabling it to move add a reasonable level of difficulty to the game.

Update the settings class to include a speed factor for the aliens.

Create method update to model the movement of the alien.

Create function update_aliens to update the positions of the fleet.

Update alien_invasion.py to update the fleet positions.