Add functionality for the fleets to change direction
In a game, the enemies would not be predictable. Static enemies are easy hits.

To make the game more challenging, the enemies should change directions.

Update Settings to add attribute fleet_direction to store direction.

Create method check_edges in Alien to check if alien is at edge of screen.

Update method update in Alien to account for changed direction.