Add functionality for dropping the fleet
The fleet of aliens are unable to approach the lower edge of the screen or the player.

This means the player will not lose the game.

As the conditions for losing are not being attempted to be met.

To fix this, the fleet should be able to move downwards.

Create function change_fleet_direction to drop the fleet and change direction.

Create function check_fleet_edges to check if fleet has reached the edge of screen.

Update function update_aliens to call change_fleet_direction and check_fleet_edges.

Update function call update_aliens in main loop to account for changes.